Mostly game logic + some tests
Prepared initial projects for FE,Game logic,Server,Tests and Shared.
Most of the game state control structures are already in place, with state machine in heavy dev.
It was a long day, and just as long back-and-forth session trying to decide on the code conventions, so expect to find some inconsistencies in the style.
Also, included are some sample tests to get us started.